test_grad/test_grad/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
torch>=2.7.1
torchvision>=0.18.1
opencv-python>=4.7.0.72
numpy>=1.24.3
Pillow>=9.5.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.1.0
